Try to avoid MSVC warning.

+++ include/xprec/ddouble.h
@@ -113,8 +113,8 @@ public:
     constexpr DDouble(long double x) : _hi(x), _lo(x - _hi) { }
 
     // TODO: requires extended precision to work.
-    constexpr DDouble(int64_t x) : _hi(x), _lo((long double)x - _hi) {}
-    constexpr DDouble(uint64_t x) : _hi(x), _lo((long double)x - _hi) {}
+    constexpr DDouble(int64_t x) : _hi((double)x), _lo((long double)x - _hi) {}
+    constexpr DDouble(uint64_t x) : _hi((double)x), _lo((long double)x - _hi) {}
     constexpr DDouble(int32_t x) : _hi(x), _lo(0.0) { }
     constexpr DDouble(uint32_t x) : _hi(x), _lo(0.0) { }
     constexpr DDouble(int16_t x) : _hi(x), _lo(0.0) { }
